#1 LoggingInterceptor interface improved
