+++ okhttplogger/src/main/java/com/drivekitt/common/utils/logging/LoggingInterceptorImpl.kt
@@ -1,0 +1,73 @@
+package com.drivekitt.common.utils.logging
+
+import okhttp3.Interceptor
+import okhttp3.Response
+import okhttp3.ResponseBody
+import java.io.IOException
+import java.util.concurrent.TimeUnit
+
+class LoggingInterceptorImpl internal constructor(private val builder: LoggingInterceptor.Builder) : LoggingInterceptor {
+
+    @Throws(IOException::class)
+    override fun intercept(chain: Interceptor.Chain): Response {
+
+        val printer = Printer(builder.loggerFactory.invoke())
+
+        val request = chain.request()
+
+        if (builder.level == LoggingInterceptor.Level.NONE) {
+            return chain.proceed(request)
+        }
+
+        val requestBody = request.body()
+
+        val rSubtype: String? = requestBody?.contentType()?.subtype()
+
+
+        if (isNotFileRequest(rSubtype)) {
+            printer.printJsonRequest(builder, request)
+        } else {
+            printer.printFileRequest(builder, request)
+        }
+
+        val st = System.nanoTime()
+        val response = chain.proceed(request)
+        val chainMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - st)
+
+        val segmentList = request.url().encodedPathSegments()
+        val header = response.headers().toString()
+        val code = response.code()
+        val isSuccessful = response.isSuccessful
+        val message = response.message()
+        val responseBody = response.body()
+        val contentType = responseBody!!.contentType()
+
+        var subtype: String? = null
+        val body: ResponseBody
+
+        if (contentType != null) {
+            subtype = contentType.subtype()
+        }
+
+        if (isNotFileRequest(subtype)) {
+            val bodyString = Printer.getJsonString(responseBody.string())
+            val url = response.request().url().toString()
+
+            printer.printJsonResponse(builder, chainMs, isSuccessful, code, header, bodyString,
+                    segmentList, message, url)
+            body = ResponseBody.create(contentType, bodyString)
+        } else {
+            printer.printFileResponse(builder, chainMs, isSuccessful, code, header, segmentList, message)
+            return response
+        }
+
+        return response.newBuilder().body(body).build()
+    }
+
+    private fun isNotFileRequest(subtype: String?): Boolean {
+        return subtype != null && (subtype.contains("json")
+                || subtype.contains("xml")
+                || subtype.contains("plain")
+                || subtype.contains("html"))
+    }
+}
